BUGFIX: Multiple playbacks on Safari now ok
It seems that on Safari, after a single playback,
the currentTime was set to the end of the file
whereas in other browsers, the currentTime resets to
zero.  Thus, in Safari, playing a second time would
result in nothing happening.

Explictly setting the time to zero seems to result
in the desired behavior for all browsers.