BUGFIX: Dataless forms correctly create entry in user data
In the past, we looked for the name in the form.  But, if
the user did not click any checkboxes, etc. then the form
would be empty.

The new code instead looks at what data is expected to be
there and uses that to serialize the data (all zero or null values
as opposed to not outputing anything).