BUGFIX: Forgot to include implicit pages, plus more
Some pages only appear as an effect of user interaction
(eg if a user_name already exists) or if the user decides
not to consent.  This code adds them in.

Also, fixed a bug where a pages with no unique text would crash
the text dictionary generator.

Also, added unicode output support.

Also, removed a spurous ')' that got printed to the screen.