FEATURE: Collapsed classes in axb_pages.py into one flexible class
AudioChoicePage()

AXB-like pages all have the same features: X number of audio buttons
and Y number of choices for the user.  This new class makes it easy
to build different kinds of pages like these, at the cost of longer
line definitions in a sequence file.  The class can be subclassed
if this is an issue.

Updated LMEDS example to reflect the new 'audio_choice' page.