BUGFIX: LMEDS secretly only used OGG files
Although it had hooks to play OGG and MP3 files, if an OGG file was
missing, it wouldn't attempt to load the MP3 file.

The user can specify what formats a test uses and all audio files
MUST be encoded in the specified formats.

The default formats are .ogg and .mp3