FEATURE: Option to let users relogin
This feature adds the page number to each line in the users output.

An often made request is to allow users to relogin to an experiment.
Under the previous system, if a user loses access to the page and tries
to log back in, they will be unable to do so--this was by design.

Now the system is configurable (in the .cgi file) to either allow users
to relogin or to deny them this ability.