BUGFIX: Audiobuttons would become prematurely active
The event listener that activated the buttons after an audio file
was finished playing was staying active.  This bug only effected
series of audio files (e.g. [water water water].  The listener was
activated at the play of the last instance of water.  But since
it wasn't removed after running, every time ANY water finished, it
was triggered again.  Now it immediately removes itself as a listner
once it plays.