BUGFIX: Autosubmit code was waiting twice as long as it needed
A timer was started at the end of the last audio file.  The timer
was the duration of the audio file.  When the timer ended, a new
timer was launched for the duration of the audio file.

The new mechanism is an event that is triggered when the audio/video
is finished playing.