BUGFIX: Post-process ok with non-uniform annotations
Before it was assumed that all annotation tasks used the same
set of instructions for all texts.  Now it is possible to have
some tasks annotated with certain instructions and others with
other instructions.

Furthermore, before it used to output all scores for the same
text horizontally.  However, now it only does this for the text
annotated under a single instruction set.

This can be viewed in the demo output
tests/lmeds_demo/output/LMEDS_Demo/duplicates_not_removed_results/