BUGFIX: Post-processing code for rpt pages was too simple
The 'unique' stimID used only a subset of the line entry in
the sequence file.  Thus, for example, if the same page
was presented with the same audio but different transcripts,
or vice-versa--incorrect data would be associated with a key.

Now the whole line is used.  It is a little more verbose, but
users can manually remove unwanted columns afterwards.