BUGFIX: -CRITICAL- bugfix for output data from randomized sequences
If a user exited prematuraly and relogged in to do an experiment
in cases where their sequence was randomized, on the second login,
they would be presented with the page appropriate if the test was
not randomized.  HOWEVER, when that data was saved, the correct
randomized page would be placed in the user's output.

In other words, the page descriptor would be for the randomized page
but the response would be for the unrandomized page.  The two
parts of the output line would not match.