FIX: LMEDS was never working with python 3
The code prints a cgi header before the html code gets printed.

However, in python 3, the html code always gets printed first, even
though the execution of the code printing happens later.  To fix,
we attach the header to the html code so that we can guarentee the
printing order